+++ src/VpnPortalModule.php
@@ -42,9 +42,6 @@ class VpnPortalModule implements ServiceModuleInterface
     /** @var bool */
     private $shuffleHosts = true;
 
-    /** @var array */
-    private $addVpnProtoPorts = [];
-
     public function __construct(TplInterface $tpl, ServerClient $serverClient, SessionInterface $session, Storage $storage, callable $getClientInfo)
     {
         $this->tpl = $tpl;
@@ -57,11 +54,6 @@ class VpnPortalModule implements ServiceModuleInterface
     public function setShuffleHosts($shuffleHosts)
     {
         $this->shuffleHosts = (bool) $shuffleHosts;
-    }
-
-    public function setAddVpnProtoPorts(array $addVpnProtoPorts)
-    {
-        $this->addVpnProtoPorts = $addVpnProtoPorts;
     }
 
     public function init(Service $service)
@@ -317,7 +309,7 @@ class VpnPortalModule implements ServiceModuleInterface
         // get the CA & tls-auth
         $serverInfo = $this->serverClient->get('server_info');
 
-        $clientConfig = ClientConfig::get($profileData, $serverInfo, $clientCertificate, $this->shuffleHosts, $this->addVpnProtoPorts);
+        $clientConfig = ClientConfig::get($profileData, $serverInfo, $clientCertificate, $this->shuffleHosts);
 
         // convert the OpenVPN file to "Windows" format, no platform cares, but
         // in Notepad on Windows it looks not so great everything on one line