+++ web/api.php
@@ -103,10 +103,6 @@ try {
         $vpnApiModule = new VpnApiModule(
             $serverClient
         );
-        if ($config->hasItem('addVpnProtoPorts')) {
-            $vpnApiModule->setAddVpnProtoPorts($config->getItem('addVpnProtoPorts'));
-        }
-
         $service->addModule($vpnApiModule);
     }
     $service->run($request)->send();