+++ web/index.php
@@ -322,10 +322,6 @@ try {
         $storage,
         $getClientInfo
     );
-    if ($config->hasItem('addVpnProtoPorts')) {
-        $vpnPortalModule->setAddVpnProtoPorts($config->getItem('addVpnProtoPorts'));
-    }
-
     $service->addModule($vpnPortalModule);
 
     // TOTP module